File reorganization
The various modules on the top level made it a bit difficult to fine one's way (and there is now yet another module). Moved the modules into a separate 'lib' folder, the only one on the top level is 'main.js'. Makes things a bit clearer...
