+++ LICENSE
@@ -1,6 +1,6 @@
 W3C SOFTWARE NOTICE AND LICENSE
 
-Copyright (c) 2017 W3C® (MIT, ERCIM, Keio, Beihang)
+Copyright (c) 2018 W3C® (MIT, ERCIM, Keio, Beihang)
 
 License
