+++ lib/convert.js
