+++ lib/io.js
