+++ lib/schemas.js
@@ -14,7 +14,7 @@ const fs  = require('fs');
 
 let get_schema = (fname) => {
     try {
-        let filename = `${__dirname}/${fname}`;
+        let filename = `${__dirname}/../${fname}`;
         let file_c = fs.readFileSync(filename, "utf-8");
         return JSON.parse(file_c);
     } catch(e) {
